requirements.txt
about-time==4.2.1
adb-shell==0.4.3
alive-progress==3.1.1
azure-common==1.1.28
azure-core==1.26.4
azure-identity==1.12.0
azure-mgmt-compute==29.1.0
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.0.0
bcrypt==4.0.1
blinker==1.6.2
boto==2.49.0
build==0.10.0
capstone==4.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
docopt==0.6.2
docutils==0.19
Flask==2.3.2
grapheme==0.6.0
hatasm @ git+https://github.com/EntySec/HatAsm@d9bbc4db0e13238227fbb7d3a99a1ecdf8bd60fd
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
keystone-engine==0.9.2
manuf==1.1.5
MarkupSafe==2.1.2
msal==1.22.0
msal-extensions==1.0.0
msrest==0.7.1
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.2
packaging==23.0
paramiko==3.1.0
pex @ git+https://github.com/EntySec/Pex@8451c431dcea655870ed2bc06ac0306c4caaf6fd
pip-tools==6.12.3
pipreqs==0.4.11
portalocker==2.7.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
PyNaCl==1.5.0
pyproject_hooks==1.0.0
python-dotenv==1.0.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scapy==2.5.0
schedule==1.2.0
six==1.16.0
slack-bolt==1.17.2
slack-sdk==3.21.2
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.3
yarg==0.1.9
